refactor: clean up config discovery internals
